requirements.txt
websockets==4.0.1
kafka-python==1.4.2